+++ src/examples/rttr_rpc_beast_server/RttRpcBeastServer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "RttRpcServiceRepository.h"
+#include "core/RttRpcServiceRepository.h"
 
 #include <boost/asio/ip/tcp.hpp>
 
@@ -8,7 +8,7 @@ class RttRpcBeastImpl;
 class RttRpcBeastServer
 {
 public:
-	RttRpcBeastServer(size_t threadCount);
+	RttRpcBeastServer(int threadCount);
 	~RttRpcBeastServer();
 	void start (const boost::asio::ip::tcp::endpoint& ep);
